spires: separation to contrib module
