+++ invenio_query_parser/walkers/pypeg_to_ast.py
@@ -24,7 +24,6 @@
 """Implement Pypeg to AST converter."""
 
 from .. import ast, parser
-from ..contrib.spires import parser as spires_parser
 from ..visitor import make_visitor
 
 
@@ -50,10 +49,6 @@ class PypegConverter(object):
         return ast.Or(left, right)
 
     @visitor(parser.KeywordRule)
-    def visit(self, node):
-        return ast.Keyword(node.value)
-
-    @visitor(spires_parser.SpiresKeywordRule)
     def visit(self, node):
         return ast.Keyword(node.value)
 
@@ -85,22 +80,6 @@ class PypegConverter(object):
     def visit(self, node, left, right):
         return ast.RangeOp(left, right)
 
-    @visitor(spires_parser.GreaterQuery)
-    def visit(self, node, child):
-        return ast.GreaterOp(child)
-
-    @visitor(spires_parser.GreaterEqualQuery)
-    def visit(self, node, child):
-        return ast.GreaterEqualOp(child)
-
-    @visitor(spires_parser.LowerQuery)
-    def visit(self, node, child):
-        return ast.LowerOp(child)
-
-    @visitor(spires_parser.LowerEqualQuery)
-    def visit(self, node, child):
-        return ast.LowerEqualOp(child)
-
     @visitor(parser.Number)
     def visit(self, node):
         return ast.Value(node.value)
@@ -113,45 +92,9 @@ class PypegConverter(object):
     def visit(self, node):
         return ast.Keyword(node.value)
 
-    @visitor(spires_parser.SpiresSimpleValue)
-    def visit(self, node):
-        return ast.Value(node.value)
-
-    @visitor(spires_parser.SpiresValue)
-    def visit(self, node, children):
-        return ast.Value("".join([c.value for c in children]))
-
-    @visitor(spires_parser.SpiresValueQuery)
-    def visit(self, node, child):
-        return ast.ValueQuery(child)
-
-    @visitor(spires_parser.SpiresSimpleQuery)
-    def visit(self, node, child):
-        return child
-
-    @visitor(spires_parser.SpiresParenthesizedQuery)
-    def visit(self, node, child):
-        return child
-
-    @visitor(spires_parser.SpiresNotQuery)
-    def visit(self, node, child):
-        return ast.AndOp(None, ast.NotOp(child))
-
-    @visitor(spires_parser.SpiresAndQuery)
-    def visit(self, node, child):
-        return ast.AndOp(None, child)
-
-    @visitor(spires_parser.SpiresOrQuery)
-    def visit(self, node, child):
-        return ast.OrOp(None, child)
-
     @visitor(parser.ValueQuery)
     def visit(self, node, child):
         return ast.ValueQuery(child)
-
-    @visitor(spires_parser.SpiresKeywordQuery)
-    def visit(self, node, keyword, value):
-        return ast.SpiresOp(keyword, value)
 
     @visitor(parser.KeywordQuery)
     def visit(self, node, keyword, value):
@@ -191,54 +134,11 @@ class PypegConverter(object):
             tree = booleanNode
         return tree
 
-    @visitor(spires_parser.SpiresQuery)
-    def visit(self, node, children):
-        # Assign implicit keyword
-        # find author x and y --> find author x and author y
-
-        def assign_implicit_keyword(implicit_keyword, node):
-            """
-            Note: this function has side effects on node content
-            """
-            if type(node) in [ast.AndOp, ast.OrOp] and \
-               type(node.right) == ast.ValueQuery:
-                node.right = ast.SpiresOp(implicit_keyword, node.right.op)
-            if type(node) in [ast.AndOp, ast.OrOp] and \
-               type(node.right) == ast.NotOp:
-                assign_implicit_keyword(implicit_keyword, node.right)
-            if type(node) in [ast.NotOp] and \
-               type(node.op) == ast.ValueQuery:
-                node.op = ast.SpiresOp(implicit_keyword, node.op.op)
-
-        implicit_keyword = None
-        for child in children:
-            new_keyword = getattr(child, 'keyword', None)
-            if new_keyword is not None:
-                implicit_keyword = new_keyword
-            if implicit_keyword is not None:
-                assign_implicit_keyword(implicit_keyword, child)
-
-        # Build the boolean expression, left to right
-        # x and y or z and ... --> ((x and y) or z) and ...
-        tree = children[0]
-        for booleanNode in children[1:]:
-            booleanNode.left = tree
-            tree = booleanNode
-        return tree
-
-    @visitor(spires_parser.FindQuery)
-    def visit(self, node, child):
-        return child
-
     @visitor(parser.EmptyQueryRule)
     def visit(self, node):
         return ast.EmptyQuery(node.value)
 
     @visitor(parser.Main)
-    def visit(self, node, child):
-        return child
-
-    @visitor(spires_parser.Main)
     def visit(self, node, child):
         return child
